Updated Sensor Information and Autonomous
Since I have last updated, I have tested the lift correction, and changed some values. In order to not have issues when driving, the potentiometers are now multiplied by a value that causes each tick to be 1 degree of movement. We have added another potentiometer on the lift, and another jumper clip for autonomous. I have also made a simple autonomous that scores up to 6 points.